Fix protocol variable types: use unsigned types
Change variable types to match the unsigned values loaded from the network
protocol, eliminating unnecessary casts and fixing sign conversion warnings:

- Change item/item_flags arrays from int to uint32_t (loaded via load_u32)
- Change originx/originy from int to uint16_t (loaded via load_u16)
- Change mirror/newmirror from int to uint32_t (loaded via load_u32)
- Change realtime from int to uint32_t (loaded via load_u32)
- Change act/actx/acty from int to uint16_t (loaded via load_u16)
- Change len variables in protocol functions from int to unsigned char or
  uint16_t (matching what's loaded from buffer)
- Change tmp32 in sv_map11 from int to uint32_t
- Change t in sv_ping/svl_ping from int to uint32_t
- Change cn in sv_name from int to uint16_t

Make protocol functions static that are only used within protocol.c:
- sv_map01, sv_map10, sv_map11
- svl_ping, sv_ping
- sv_scroll_* functions
- sv_setval, sv_sethp, sv_endurance, sv_lifeshield, sv_setmana, sv_setrage
- sv_setitem, sv_setorigin, sv_settick
- sv_mirror, sv_realtime, sv_speedmode, sv_fightmode
- sv_setcitem, sv_act
- sv_text, svl_text, sv_conname, svl_conname
- sv_exit, svl_exit, sv_name

This addresses compiler warnings by fixing the root cause (incorrect types for
unsigned protocol values) rather than adding casts, ensuring type safety and
preventing potential bugs from sign mismatches.

+++ src/client/client.c
@@ -42,11 +42,11 @@ static struct z_stream_s zs;
 
 DLL_EXPORT char username[40];
 DLL_EXPORT uint32_t tick;
-DLL_EXPORT int mirror = 0;
-DLL_EXPORT int realtime;
+DLL_EXPORT uint32_t mirror = 0;
+DLL_EXPORT uint32_t realtime;
 DLL_EXPORT int protocol_version = 0;
 
-int newmirror = 0;
+uint32_t newmirror = 0;
 int lasttick; // ticks in inbuf
 static size_t lastticksize; // size inbuf must reach to get the last tick complete in the queue
 
@@ -64,21 +64,21 @@ static unsigned char inbuf[MAX_INBUF];
 static size_t outused;
 static unsigned char outbuf[MAX_OUTBUF];
 
-DLL_EXPORT int act;
-DLL_EXPORT int actx;
-DLL_EXPORT int acty;
+DLL_EXPORT uint16_t act;
+DLL_EXPORT uint16_t actx;
+DLL_EXPORT uint16_t acty;
 
 DLL_EXPORT unsigned int cflags; // current item flags
 DLL_EXPORT unsigned int csprite; // and sprite
 
-DLL_EXPORT int originx;
-DLL_EXPORT int originy;
+DLL_EXPORT uint16_t originx;
+DLL_EXPORT uint16_t originy;
 DLL_EXPORT struct map map[MAPDX * MAPDY];
 DLL_EXPORT struct map map2[MAPDX * MAPDY];
 
 DLL_EXPORT int value[2][V_MAX];
-DLL_EXPORT int item[INVENTORYSIZE];
-DLL_EXPORT int item_flags[INVENTORYSIZE];
+DLL_EXPORT uint32_t item[INVENTORYSIZE];
+DLL_EXPORT uint32_t item_flags[INVENTORYSIZE];
 DLL_EXPORT int hp;
 DLL_EXPORT int mana;
 DLL_EXPORT int rage;
